Move propagateOrbits to propagate directory
