+++ thor/orbits/__init__.py
@@ -1,3 +1,2 @@
 from .kepler import *
-from .pyoorb import *
 from .propagate import *
