+++ thor/orbits/propagate/__init__.py
@@ -1,2 +1,3 @@
 from .stumpff import *
 from .universal import *
+from .pyoorb import *
